Move vector tests to plugin module and fix remaining references
- Move 5 vector test files from pinot-core to pinot-plugins/pinot-vector
- Update package declarations and imports in test files
- Add mockito test dependency to pinot-vector pom.xml
- Remove FilterKind.VECTOR_SIMILARITY from FunctionRegistryTest
