+++ pinot-plugins/pinot-vector/pom.xml
@@ -51,5 +51,10 @@
       <artifactId>pinot-segment-spi</artifactId>
       <scope>provided</scope>
     </dependency>
+    <dependency>
+      <groupId>org.mockito</groupId>
+      <artifactId>mockito-core</artifactId>
+      <scope>test</scope>
+    </dependency>
   </dependencies>
 </project>
